// lib/shared/src/models/tags.ts
export enum ModelTag {
  Enterprise = 'enterprise',
  Pro = 'pro',
  Free = 'free',
  Waitlist = 'waitlist',
  Custom = 'custom',
  Recommended = 'recommended',
  Deprecated = 'deprecated',
  Experimental = 'experimental',
  Internal = 'internal',
  Vision = 'vision',
  Reasoning = 'reasoning',
  Power = 'power',
  StreamDisabled = 'stream-disabled',
  Accuracy = 'accuracy',
  Speed = 'speed',
  Balancer = 'balancer',
}